Implemented support for attribute arrays in addition to hashes.
The order of XML attributes should not matter.  However, for reproducibility
and to support deviations from XML standard, Fern will support an array of
attributes.

Amusing note: xhtml is *not* XML.  The tags are rendered by the browser in the
order they are given.  Example:

  <div>
    <span>One</span>
    <span>Two</span>
    <span>Three</span>
  </div>

If this were rendered as proper XML, the three spans would render in random
order.  It would require a different browser-supported approach, such as:

  <div>
    <span order="1">One</span>
    <span order="2">Two</span>
    <span order="3">Three</span>
  </div>